Always return WebDriverElement from Navigator::findElement()
