+++ composer.json
@@ -38,6 +38,8 @@
     "require-dev": {
         "phpunit/phpunit": "~8.0",
         "squizlabs/php_codesniffer": "3.*",
-        "phpstan/phpstan": "^0.11.15"
+        "phpstan/phpstan": "^0.11.15",
+        "mockery/mockery": "^1.2",
+        "phpstan/phpstan-mockery": "^0.11.3"
     }
 }
