+++ phpstan.neon
@@ -1,0 +1,2 @@
+includes:
+  - vendor/phpstan/phpstan-mockery/extension.neon
