+++ tests/Unit/NavigatorTest.php
@@ -1,0 +1,41 @@
+<?php
+/** @noinspection PhpUnhandledExceptionInspection */
+
+namespace webignition\SymfonyDomCrawlerNavigator\Tests\Unit;
+
+use Facebook\WebDriver\WebDriver;
+use Symfony\Component\Panther\DomCrawler\Crawler;
+use webignition\SymfonyDomCrawlerNavigator\CrawlerFactory;
+use webignition\SymfonyDomCrawlerNavigator\Exception\UnknownElementException;
+use webignition\SymfonyDomCrawlerNavigator\Model\ElementLocator;
+use webignition\SymfonyDomCrawlerNavigator\Model\LocatorType;
+use webignition\SymfonyDomCrawlerNavigator\Navigator;
+
+class NavigatorTest extends \PHPUnit\Framework\TestCase
+{
+    public function testFindElementThrowsUnknownElementException()
+    {
+        $crawler = new Crawler([], \Mockery::mock(WebDriver::class));
+        $elementCrawler = new Crawler([], \Mockery::mock(WebDriver::class));
+
+        $crawlerFactory = \Mockery::mock(CrawlerFactory::class);
+        $crawlerFactory
+            ->shouldReceive('createElementCrawler')
+            ->andReturn($elementCrawler);
+
+        $navigator = new Navigator($crawler, $crawlerFactory);
+
+        $elementLocator = new ElementLocator(
+            LocatorType::CSS_SELECTOR,
+            '.does-not-exist',
+            1
+        );
+
+        try {
+            $navigator->findElement($elementLocator);
+            $this->fail('UnknownElementException not thrown');
+        } catch (UnknownElementException $unknownElementException) {
+            $this->assertSame($elementLocator, $unknownElementException->getElementLocator());
+        }
+    }
+}
